Added code for scraping pdfs from Google Scholar

+++ question_answering/inference.py
@@ -19,36 +19,40 @@ def inference_with_verifiers(search_query, decomposition, facts_text, key_questi
     MAX_TOKENS = 10000
     model = "gpt-3.5-turbo"
     search_query_file_safe = sanitize_filename(search_query)
-
-    # Get initial inferences
-    initial_inference = chat_openai(get_initial_inference_prompt(decomposition, facts_text, MAX_TOKENS), model=model)[0]
-
-    # Get verifier feedback on initial inference
-    inference_verifiers_res = chat_openai(get_inference_verifiers_prompt(initial_inference), model=model)[0]
-
-    # Improve original decomposition with verifier feedback
-    improved_inference_res = chat_openai(get_inference_improved_prompt(initial_inference, inference_verifiers_res), model=model)[0]
-
-    # Get verifier feedback on final decomposition
-    improved_inference_verifiers_res = chat_openai(get_inference_verifiers_prompt(improved_inference_res), model=model)[0]
-
-    improved_inference_res_json = json.loads(improved_inference_res)
-
     # Save decomposition and verifier feedback to file
     if not os.path.exists(f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}'):
         os.makedirs(f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}')
 
-    with open(f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/initial_inference.json', 'w') as f:
-        json.dump(json.loads(initial_inference), f, indent=2)
-    with open(f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/initial_verifier_feedback.txt', 'w', encoding='utf-8') as f:
-        f.write(inference_verifiers_res)
-    with open(f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/improved_inference.json', 'w') as f:
-        json.dump(improved_inference_res_json, f, indent=2)
-    with open(f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/improved_verifier_feedback.txt', 'w', encoding='utf-8') as f:
-        f.write(improved_inference_verifiers_res)
+    # Get initial inferences
+    save_path = f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/initial_inference.json'
+    if not os.path.exists(save_path):
+        initial_inference = chat_openai(get_initial_inference_prompt(decomposition, facts_text, MAX_TOKENS), model=model)[0]
+        with open(save_path, 'w') as f:
+            json.dump(json.loads(initial_inference), f, indent=2)
 
-    # Return final inference, along with verifier feedback
-    return improved_inference_res_json, improved_inference_verifiers_res
+    # Get verifier feedback on initial inference
+    save_path = f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/initial_verifier_feedback.txt'
+    if not os.path.exists(save_path):
+        inference_verifiers_res = chat_openai(get_inference_verifiers_prompt(initial_inference), model=model)[0]
+        with open(save_path, 'w', encoding='utf-8') as f:
+            f.write(inference_verifiers_res)
+
+    # Improve original decomposition with verifier feedback
+    save_path = f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/improved_inference.json'
+    if not os.path.exists(save_path):
+        improved_inference_res = chat_openai(get_inference_improved_prompt(initial_inference, inference_verifiers_res), model=model)[0]
+        with open(save_path, 'w') as f:
+            json.dump(json.loads(improved_inference_res), f, indent=2)
+
+    # Get verifier feedback on final decomposition
+    save_path = f'autoscious_logs/{search_query_file_safe}/inferences/kq{key_question_idx}/improved_verifier_feedback.txt'
+    if not os.path.exists(save_path):
+        improved_inference_verifiers_res = chat_openai(get_inference_verifiers_prompt(improved_inference_res), model=model)[0]
+        with open(save_path, 'w', encoding='utf-8') as f:
+            f.write(improved_inference_verifiers_res)
+
+    # Return because all content should be in files
+    return
 
 def get_hypothesis_inference_findings_from_facts(search_query):
     # Load in decomposition
@@ -81,9 +85,6 @@ def get_hypothesis_inference_findings_from_facts(search_query):
     for i, kq_decomposition in enumerate(key_question_decomposition_list):
         with open(f'autoscious_logs/{sanitize_filename(search_query)}/facts/kq{i}/facts.txt', 'r', encoding='utf-8') as f:
             kq_facts = f.read()
-        inference, inference_verifier_feedback = inference_with_verifiers(search_query, kq_decomposition, kq_facts, i)
-
-        print("\n Inference: \n", inference)
-        print("\n Inference verifier feedback: \n", inference_verifier_feedback)
+        inference_with_verifiers(search_query, kq_decomposition, kq_facts, i)
 
 get_hypothesis_inference_findings_from_facts(search_query)
